==> 1_combine_each_dataset.ipynb <==
import pandas as pd
import numpy as np
import os

INITIAL_FOLDER = "initial_datasets"
COMBINED_FOLDER = "combined_datasets"
SCOPE = "train_val"
DATASET = "django_django"

columns = [
    "number","title","created_at","closed_at","merged_at","additions","deletions","changed_files",
    "commits","author","comments_list","reviewers_list","commits_list","files_metrics","min_max_nesting",
    "avg_max_nesting","max_max_nesting","min_func_count","avg_func_count","max_func_count","min_max_args",
    "avg_max_args","max_max_args","min_call_count","avg_call_count","max_call_count","min_if_count",
    "avg_if_count","max_if_count","min_loop_count","avg_loop_count","max_loop_count","min_avg_cc",
    "avg_avg_cc","max_avg_cc","min_max_cc","avg_max_cc","max_max_cc","min_loc","avg_loc","max_loc",
    "min_lloc","avg_lloc","max_lloc","min_sloc","avg_sloc","max_sloc","min_comments","avg_comments",
    "max_comments","min_multi_comments","avg_multi_comments","max_multi_comments","min_blank","avg_blank",
    "max_blank","title_length","description_length","files_with_content","is_bugfix","is_refactor","is_feature"
]

csv_files = []
for i in range (1,159):
    csv_files.append(f"{INITIAL_FOLDER}/{SCOPE}/{DATASET}/{DATASET}_initial_{i}.csv")
print(len(csv_files))

dfs = []
for f in csv_files:
    df = pd.read_csv(f, header=None)
    first_row = df.iloc[0].tolist()
    if first_row == columns:
        df = df.iloc[1:]
    df.columns = columns
    dfs.append(df)

combined_df = pd.concat(dfs, ignore_index=True)
combined_df = combined_df.drop_duplicates(subset="number", keep="first")
os.makedirs("final_datasets", exist_ok=True)
combined_df.to_csv(f"{COMBINED_FOLDER}/{SCOPE}/{DATASET}_combined.csv", index=False)
print(f"Combined CSV saved with {len(combined_df)} rows")

combined_df
