==> 2_combine_all_datasets.ipynb <==
import pandas as pd
import numpy as np
import os

import plt as plt

COMBINED_FOLDER = "combined_datasets"

SCOPE = "train_val"
csv_files = [f"{COMBINED_FOLDER}/{SCOPE}/django_django_combined.csv",
             f"{COMBINED_FOLDER}/{SCOPE}/keras-team_keras_combined.csv",
             f"{COMBINED_FOLDER}/{SCOPE}/fastai_fastai_combined.csv",
             f"{COMBINED_FOLDER}/{SCOPE}/matplotlib_matplotlib_combined.csv",
             f"{COMBINED_FOLDER}/{SCOPE}/scrapy_scrapy_combined.csv",
             f"{COMBINED_FOLDER}/{SCOPE}/sqlalchemy_sqlalchemy_combined.csv",
             f"{COMBINED_FOLDER}/{SCOPE}/tiangolo_fastapi_combined.csv"]

"""SCOPE = "test"
csv_files = [f"{COMBINED_FOLDER}/{SCOPE}/numpy_numpy_combined.csv",
             f"{COMBINED_FOLDER}/{SCOPE}/pallets_flask_combined.csv",
             f"{COMBINED_FOLDER}/{SCOPE}/pandas-dev_pandas_combined.csv",
             f"{COMBINED_FOLDER}/{SCOPE}/psf_requests_combined.csv",
             f"{COMBINED_FOLDER}/{SCOPE}/scikit-learn_scikit-learn_combined.csv"]"""

dfs = []
for f in csv_files:
    df = pd.read_csv(f)
    dfs.append(df)

df = pd.concat(dfs, ignore_index=True)
df.to_csv(f"{COMBINED_FOLDER}/{SCOPE}/combined.csv", index=False)
print(f"Combined CSV saved with {len(df)} rows")

print("Shape:", df.shape)
print("Columns:", df.columns.tolist())

df
