==> 5_create_final_datasets.ipynb <==
import pandas as pd

AUGMENTED_CSV = f"final_datasets/test/augmented.csv"
FINAL_CSV = f"final_datasets/test/final.csv"

df = pd.read_csv(AUGMENTED_CSV)
df

df = df.drop(columns=[c for c in df.columns.tolist() if c not in ['author_acceptance_rate', 'core_contributor_flag', 'previous_prs', 'year_created', 'time_to_close', 'comments_burstiness', 'num_comments', 'num_reviewers', 'avg_max_args', 'avg_multi_comments', 'max_max_cc', 'min_call_count', 'pr_time_label']])

df.replace(-1, pd.NA, inplace=True)
df.dropna(inplace=True)
df.reset_index(drop=True, inplace=True)
print(f"Rows: {len(df)}  Columns: {len(df.columns)}")

df['pr_time_label'] = df['pr_time_label'].replace({'immediate': 'accepted', 'fast': 'accepted', 'moderate': 'accepted', 'long': 'accepted', 'never': 'rejected'})
print(df['pr_time_label'].unique())
df['pr_time_label'].value_counts()

numeric_df = df.select_dtypes(include=['number'])
set(df.columns) - set(numeric_df.columns)

# df['additions'] = pd.to_numeric(df['additions'], errors='coerce')
# df['deletions'] = pd.to_numeric(df['deletions'], errors='coerce')

numeric_df = df.select_dtypes(include=['number'])
set(df.columns) - set(numeric_df.columns)

df.to_csv(FINAL_CSV, index=False)

df
