==> 6_train_val_test.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.metrics import classification_report, confusion_matrix, accuracy_score, f1_score
from sklearn.preprocessing import StandardScaler
from sklearn.utils import class_weight
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense, Dropout, BatchNormalization
from sklearn.ensemble import RandomForestClassifier
from xgboost import XGBClassifier
from catboost import CatBoostClassifier
from pytorch_tabnet.tab_model import TabNetClassifier
import torch

TRAIN_VAL_DATASET = "final_datasets/train_val/final.csv"
TEST_DATASET = "final_datasets/test/final.csv"

df_train_val = pd.read_csv(TRAIN_VAL_DATASET)
df_test = pd.read_csv(TEST_DATASET)

common = pd.merge(df_train_val, df_test, how='inner')
print(common)
print("Number of identical rows:", len(common))

df_train_val.round()
df_train_val

print(df_train_val.columns.tolist())

df_test = df_test[df_train_val.columns]
df_test.round()
df_test

common = pd.merge(df_train_val, df_test, how='inner')
print(common)
print("Number of identical rows:", len(common))

label_values = sorted(df_train_val['pr_time_label'].unique(), reverse = True)
label_map = {label: i for i, label in enumerate(label_values)}
print("LABEL MAP =", label_map)

X = df_train_val.drop("pr_time_label", axis=1)
y = df_train_val['pr_time_label'].map(label_map).values
print("Encoded labels:", np.unique(y))

X_train, X_val, y_train, y_val = train_test_split(X, y, test_size=0.20, random_state=42, stratify=y)
print(f"Train size: {len(X_train)}, Validation size: {len(X_val)}")
print("Distribution train: ", np.unique(y_train, return_counts=True))
print("Distribution val: ", np.unique(y_val, return_counts=True))

class_weights = class_weight.compute_class_weight(
    class_weight='balanced',
    classes=np.unique(y_train),
    y=y_train
)
class_weights_dict = dict(enumerate(class_weights))
print("Class weights:", class_weights_dict)

X_test = df_test.drop("pr_time_label", axis=1)
y_test = df_test['pr_time_label'].map(label_map).values
print("Encoded labels:", np.unique(y_test))
print("Distribution test: ", np.unique(y_test, return_counts=True))

rf = RandomForestClassifier(
    n_estimators=1000,
    max_depth=15,
    min_samples_split=50,
    min_samples_leaf=50,
    max_features='sqrt',
    bootstrap=True,
    random_state=42,
    n_jobs=-1
)

rf.fit(X_train, y_train)

rf_pred_train = rf.predict(X_train)
print("Training Accuracy:", accuracy_score(y_train, rf_pred_train))
print(classification_report(y_train, rf_pred_train, digits=4))

rf_pred_val = rf.predict(X_val)
print("Validation Accuracy:", accuracy_score(y_val, rf_pred_val))
print(classification_report(y_val, rf_pred_val, digits=4))

rf_pred_test = rf.predict(X_test)
print("Testing Accuracy:", accuracy_score(y_test, rf_pred_test))
print(classification_report(y_test, rf_pred_test, digits=4))

xgb = XGBClassifier(
    n_estimators=330,       # OK fewer trees for stability
    max_depth=2,            # OK slightly deeper than before
    learning_rate=0.118,    # OK slightly higher to learn faster
    subsample=0.8,          # OK row sampling
    colsample_bytree=0.9,   # OK feature sampling
    gamma=0.1,              # OK minimum loss reduction
    reg_alpha=0.05,         # OK mild L1 regularization
    reg_lambda=1.0,         # OK L2 regularization
    objective='binary:logistic',
    eval_metric='logloss',
    tree_method='hist',
    n_jobs=-1,
    random_state=42
)

xgb.fit(
    X_train, y_train,
    eval_set=[(X_val, y_val)],
    verbose=50
)

xgb_pred_train = xgb.predict(X_train)
print("Training Accuracy:", accuracy_score(y_train, xgb_pred_train))
print(classification_report(y_train, xgb_pred_train, digits=4))

xgb_pred_val = xgb.predict(X_val)
print("Validation Accuracy:", accuracy_score(y_val, xgb_pred_val))
print(classification_report(y_val, xgb_pred_val, digits=4))

xgb_pred_test = xgb.predict(X_test)
print("Testing Accuracy:", accuracy_score(y_test, xgb_pred_test))
print(classification_report(y_test, xgb_pred_test, digits=4))

y_proba = xgb.predict_proba(X_test)[:,1]
threshold = 0.58
y_pred = (y_proba >= threshold).astype(int)
print(classification_report(y_test, y_pred, digits=4))

cat = CatBoostClassifier(
    iterations=10000,
    learning_rate=0.01,
    depth=1,
    eval_metric='Accuracy',
    random_seed=42,
    verbose=50
)

cat.fit(X_train, y_train, eval_set=(X_val, y_val))

cat_pred_train = cat.predict(X_train)
print("Training Accuracy:", accuracy_score(y_train, cat_pred_train))
print(classification_report(y_train, cat_pred_train, digits=4))

cat_pred_val = cat.predict(X_val)
print("Validation Accuracy:", accuracy_score(y_val, cat_pred_val))
print(classification_report(y_val, cat_pred_val, digits=4))

cat_pred_test = cat.predict(X_test)
print("Testing Accuracy:", accuracy_score(y_test, cat_pred_test))
print(classification_report(y_test, cat_pred_test, digits=4))

X_train = X_train.to_numpy()
X_val = X_val.to_numpy()
X_test = X_test.to_numpy()

clf = TabNetClassifier(
    n_d=128, n_a=128,
    n_steps=20,
    gamma=1.5,
    lambda_sparse=1e-4,
    mask_type="entmax",

    optimizer_fn=torch.optim.AdamW,
    optimizer_params=dict(lr=0.009, weight_decay=1e-4),

    scheduler_fn=torch.optim.lr_scheduler.CosineAnnealingLR,
    scheduler_params={"T_max":50, "eta_min":1e-5}
)

clf.fit(
    X_train,
    y_train,
    eval_set=[(X_train, y_train), (X_val, y_val)],
    eval_name=['train', 'test'],
    eval_metric=['accuracy', 'auc'],
    max_epochs=20,
    batch_size=128,
    virtual_batch_size=64,
    patience=50,
    drop_last=False
)

tab_pred_train = clf.predict(X_train)
print("Training Accuracy:", accuracy_score(y_train, tab_pred_train))
print(classification_report(y_train, tab_pred_train, digits=4))

tab_pred_val = clf.predict(X_val)
print("Validation Accuracy:", accuracy_score(y_val, tab_pred_val))
print(classification_report(y_val, tab_pred_val, digits=4))

tab_pred_test = clf.predict(X_test)
print("Testing Accuracy:", accuracy_score(y_test, tab_pred_test))
print(classification_report(y_test, tab_pred_test, digits=4))
